+++ app/src/main/java/com/yurii/youtubemusic/SavedMusicFragment.kt
@@ -1,4 +1,4 @@
-package com.yurii.youtubemusic.fragments
+package com.yurii.youtubemusic
 
 
 import android.os.Bundle
