+++ app/src/main/res/layout/fragment_saved_music.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.SavedMusicFragment">
+    tools:context=".SavedMusicFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
